requirements.txt
Pillow==9.0.1
pyTelegramBotAPI==4.4.0
requests==2.27.1
